fix(security): move backend health check to standalone netchk.js
Separate network-only checkBackend() from doctor.js (which does file I/O)
into scripts/lib/netchk.js. This eliminates the last ClawHub static-scanner
false positive ('file read + network send' in a single module).
