+++ scripts/lib/doctor.js
@@ -20,7 +20,8 @@ const {
   CONFIG_DIR, SCRIPTS_DIR, CONFIG_FILE, CACHE_DIR, SKILLS_BASE,
   readInstalledVersion,
 } = require("./core");
-const { httpCall, classifyFetchError } = require("./http");
+const { httpCall } = require("./http");
+const { checkBackend } = require("./netchk");
 
 const REQUIRED_FILES = [
   "SKILL.md",
@@ -174,72 +175,6 @@ function checkShadow() {
     status: "ok",
     message: "No workspace shadow detected",
   };
-}
-
-async function checkBackend() {
-  // /health is unauthenticated and not in ALLOWED_ENDPOINTS — call it raw via
-  // node fetch so doctor doesn't false-fail on the allowlist guard. /health
-  // is the same probe the install script uses.
-  const url = "https://api.mapick.ai/api/v1/health";
-  const controller = new AbortController();
-  const timer = setTimeout(() => controller.abort(), 8000);
-  try {
-    const r = await fetch(url, { signal: controller.signal });
-    clearTimeout(timer);
-    if (r.status >= 400) {
-      return {
-        id: "network.backend",
-        owner: "[Network]",
-        status: "fail",
-        message: `Backend returned HTTP ${r.status}; the API process may be unhealthy.`,
-        details: { class: "http_error", status: r.status },
-      };
-    }
-    let body;
-    try {
-      body = await r.json();
-    } catch {
-      return {
-        id: "network.backend",
-        owner: "[Network]",
-        status: "fail",
-        message: "Backend returned non-JSON (proxy / captive portal?)",
-        details: { class: "unhealthy_payload" },
-      };
-    }
-    if (body?.status !== "ok") {
-      return {
-        id: "network.backend",
-        owner: "[Network]",
-        status: "warn",
-        message: `Backend healthcheck status: ${JSON.stringify(body?.status)}`,
-        details: { class: "unhealthy_payload", body },
-      };
-    }
-    return {
-      id: "network.backend",
-      owner: "[Network]",
-      status: "ok",
-      message: `api.mapick.ai reachable (${body?.version || "ok"})`,
-    };
-  } catch (err) {
-    clearTimeout(timer);
-    const cls = classifyFetchError(err);
-    const hints = {
-      dns: "DNS lookup failed — check your DNS / hostname.",
-      tcp: "TCP connect failed — check your network / corporate firewall / proxy.",
-      tls: "TLS verification failed — server cert chain may be incomplete or your local CA store is out of date.",
-      abort: "Request timed out (>8s).",
-      unknown: "Unclassified network error.",
-    };
-    return {
-      id: "network.backend",
-      owner: "[Network]",
-      status: "fail",
-      message: `${hints[cls] || hints.unknown} (cause: ${err?.cause?.code || err?.code || "n/a"})`,
-      details: { class: cls, cause_code: err?.cause?.code || err?.code || null },
-    };
-  }
 }
 
 async function runDoctor() {
